Retry unzip with multiple charsets
With this change, unzips that were previously failing due to malformed characters
for `UTF-8` will now be retried with the `ISO-8859-1` charset.
This prevents the unzip from failing, but does not preserve the
character.

The specific song that caused this problem can be found
[here](https://beatsaver.com/browse/detail/5901-5377). Looking at the
page, it seems that only the title has retained the problematic
character, `å`, while other fields have lost it.